Change the admin homepage structure to render packages in place
