+++ admin-portal/package.json
@@ -16,8 +16,8 @@
     "react-dom": "^18.1.0",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
-    "recharts": "^2.1.10",
     "react-table": "^7.8.0",
+    "recharts": "^2.1.10",
     "typescript": "^4.7.3",
     "web-vitals": "^2.1.4"
   },
