+++ admin-portal/src/pages/home/home.tsx
@@ -1,25 +1,33 @@
-import React, { ReactElement, useEffect, useState } from "react";
-import { useNavigate } from "react-router-dom";
+import React, { useEffect, useState } from "react";
 import { PageSelection } from "../../types/pages";
 import { Page } from "../../layout/page";
-import { AidPackages } from "../../types/AidPackages";
-import { TableRows } from "./tableRows/tableRows";
 import "./home.css";
 import { AidPackageService } from "../../apis/services/AidPackageService";
+import {AidPackage} from "../../types/AidPackage";
+import {useNavigate} from "react-router-dom";
 
 export function Home() {
-  const [aidPackages, setAidPackages] = useState<AidPackages>();
+  const [aidPackages, setAidPackages] = useState<AidPackage[]>([]);
+  const navigate = useNavigate();
 
   useEffect(() => {
     const fetchData = async () => {
       const response = await AidPackageService.getAidPackages();
       if (response.data) {
-        setAidPackages({ aidPackages: response.data });
+        setAidPackages(response.data);
       }
     };
 
     fetchData().catch(console.error);
   }, []);
+
+  function handlePledgesButton(packageId: number) {
+    navigate(`/packages/${packageId}`)
+  }
+
+  function handleDetailsButton(packageId: number) {
+    navigate(`/packages/${packageId}/pledge-status`)
+  }
 
   return (
     <Page selection={PageSelection.HOME}>
@@ -42,7 +50,36 @@ export function Home() {
               <th>Supplier</th>
               <th></th>
             </tr>
-            <TableRows aidPackages={aidPackages} />
+            {aidPackages.length == 0 && (
+                <div>No data found</div>
+            )}
+            {aidPackages.map((aidPackage) => {
+              return (
+                  <tr>
+                    <td>{aidPackage.name}</td>
+                    <td>{aidPackage.status}</td>
+                    <td>
+                      {(
+                          (aidPackage.receivedAmount / aidPackage.goalAmount || 0) * 100
+                      ).toFixed(0)}
+                      %
+                    </td>
+                    <td>--</td>
+                    <td>
+                      <button
+                          onClick={() => handleDetailsButton(aidPackage.packageID)}
+                      >
+                        Details
+                      </button>
+                      <button
+                          onClick={() => handlePledgesButton(aidPackage.packageID)}
+                      >
+                        Pledges
+                      </button>
+                    </td>
+                  </tr>
+              );
+            })}
           </table>
         </div>
       </div>
