--- admin-portal/src/pages/home/tableRow/emptyRow.tsx
@@ -1,7 +1,0 @@
-import React, { ReactElement, useEffect, useState } from "react";
-
-interface EmptyRowProps {}
-
-export function EmptyRow(props: EmptyRowProps) {
-  return <tr>No Data Found</tr>;
-}
